Added HMMM conversion, misc updates
 - Added several validation PSFs
 - Added CGUIBrowserProcess.get_jobid() (CGBP.get_jobid())
 - Initial step uses a lock to prevent duplicate jobids; this results in
   significant slowdown for some modules, but is the only *easy* way to
   implement this without manually locking in each CGBP subclass, which
   I don't trust others to do properly. Locking is handled by CGBP
   directly.
 - CGBP.go_next() allows specifying next_button element
 - renamed "find_by" to be consistent with splinter's naming
 - updated several __init__ functions' headers to make changing CGBP
   __init__ signature less painful for subclasses
 - added PDB files missing from some test cases, which had previously
   only downloaded from RCSB
 - clarified wording of some validation errors

+++ save_references.py
@@ -67,6 +67,10 @@ with args.results_file as fh:
         else:
             dest = os.path.join('files', 'references', module, ref_filename)
 
+        if not os.path.exists(src):
+            print("Skipping nonexistent file:", src)
+            continue
+
         print('copying', src, '->', dest)
         shutil.copy(src, dest)
